+++ src/App.jsx
@@ -2,15 +2,15 @@ import React from 'react';
 import {ThemeProvider} from 'styled-components';
 import {BrowserRouter as Router, Routes, Route} from 'react-router-dom';
 
-import Header from 'component/common/Header';
-import ProductListPage from 'page/ProductListPage';
-import ProductDetailPage from 'page/ProductDetailPage';
-import ProductCartPage from 'page/ProductCartPage';
+import Header from 'components/common/Header';
+import ProductListPage from 'pages/ProductListPage';
+import ProductDetailPage from 'pages/ProductDetailPage';
+import ProductCartPage from 'pages/ProductCartPage';
 
 import theme from 'theme/theme';
 import {GlobalStyles} from 'style';
 
-import {PATH} from 'constant';
+import {PATH} from 'constants';
 
 export default function App() {
   return (
