--- src/component/DetailItem/index.stories.jsx
@@ -1,17 +1,0 @@
-import React from 'react';
-
-import DetailItem from 'component/DetailItem';
-
-export default {
-  component: DetailItem,
-  title: 'DetailItem',
-};
-
-const Template = (args) => <DetailItem {...args} />;
-export const Defaults = Template.bind({});
-Defaults.args = {
-  itemImgURL:
-    'https://upload.wikimedia.org/wikipedia/commons/thumb/a/a7/React-icon.svg/1024px-React-icon.svg.png',
-  itemName: 'PET보틀-정사각(420ml)',
-  itemPrice: '43,400',
-};
