--- src/component/Loader/index.jsx
@@ -1,6 +1,0 @@
-import React from 'react';
-import {LoaderWrapper} from 'component/Loader/style';
-
-export default function Loading() {
-  return <LoaderWrapper />;
-}
