--- src/component/common/CheckBox/index.stories.jsx
@@ -1,12 +1,0 @@
-import React from 'react';
-
-import CheckBox from 'component/common/CheckBox';
-
-export default {
-  component: CheckBox,
-  title: 'CheckBox',
-};
-
-const Template = (args) => <CheckBox {...args} />;
-
-export const Defaults = Template.bind({});
