--- src/component/common/Header/index.stories.jsx
@@ -1,17 +1,0 @@
-import React from 'react';
-import {BrowserRouter} from 'react-router-dom';
-
-import Header from 'component/common/Header';
-
-export default {
-  component: Header,
-  title: 'Header',
-};
-
-const Template = (args) => (
-  <BrowserRouter>
-    <Header {...args} />
-  </BrowserRouter>
-);
-
-export const Defaults = Template.bind({});
