+++ src/components/DetailItem/style.js
