+++ src/components/Loader/index.jsx
@@ -1,0 +1,6 @@
+import React from 'react';
+import {LoaderWrapper} from 'components/Loader/style';
+
+export default function Loading() {
+  return <LoaderWrapper />;
+}
