+++ src/components/Loader/style.js
