+++ src/components/common/CheckBox/index.stories.jsx
@@ -1,0 +1,12 @@
+import React from 'react';
+
+import CheckBox from 'components/common/CheckBox';
+
+export default {
+  component: CheckBox,
+  title: 'CheckBox',
+};
+
+const Template = (args) => <CheckBox {...args} />;
+
+export const Defaults = Template.bind({});
