+++ src/components/common/CheckBox/style.js
